==> notebooks/Jingyi_notebooks/5_misc_doc.ipynb <==
import pandas as pd
import numpy as np
from plotnine import *
import matplotlib.pyplot as plt

df = pd.read_parquet("multi_model_1000_0.9_10000_100_RMSE_0.95_600_1/results.parquet")

df

results = df.groupby(["hyper_idx", "h"]).agg(
        mean_squared_error=("squared_error", "mean"),
        RMSE=("squared_error", lambda x: np.sqrt(x.mean())),
        MAD=("squared_error", lambda x: np.median(np.sqrt(x))),
        coverage_accuracy=("coverage_flags", "mean"),
        sd = ("sd", "mean")
    )

results["normalized_rmse"] = results["RMSE"] / results["sd"]

results.drop(columns="sd", inplace=True)

results

results.groupby(level="hyper_idx").diff()
